Move sidebar field selects into the sidebar folder and extract SelectAssignee
